+++ packages/feedback-react/documentation/FeedbackExample.tsx
@@ -63,6 +63,7 @@ export const FeedbackExample: VFC<ExampleComponentProps> = ({ boolValues, choice
             onSubmit={console.log}
             followup={followup}
             contactQuestion={contact}
+            maxLength={250}
         />
     );
 };
@@ -117,5 +118,6 @@ import { Feedback, PRESETS } from "@fremtind/jkl-feedback-react";
     }`
             : undefined
     }}
+    maxLength={250}
 />
 `;